Use timeservice for dates and move date setting to appearance tab
