+++ src/main/java/com/faforever/client/preferences/LocalizationPrefs.java
@@ -9,9 +9,11 @@ import java.util.Locale;
 
 public class LocalizationPrefs {
   private final ObjectProperty<Locale> language;
+  private final ObjectProperty<DateInfo> dateFormat;
 
   public LocalizationPrefs() {
     language = new SimpleObjectProperty<>();
+    dateFormat = new SimpleObjectProperty<>(DateInfo.AUTO);
   }
 
   @Nullable
@@ -25,5 +27,13 @@ public class LocalizationPrefs {
 
   public ObjectProperty<Locale> languageProperty() {
     return language;
+  }
+
+  public DateInfo getDateFormat() {
+    return dateFormat.get();
+  }
+
+  public void setDateFormat(DateInfo date) {
+    this.dateFormat.set(date);
   }
 }
