+++ src/main/resources/i18n/messages.properties
@@ -854,6 +854,8 @@ game.generateMap.none = None
 game.generateMap.lots = Lots
 settings.appearance.backgroundImage = Background image
 settings.appearance.backgroundImage.description = Setting a background image will increase memory consumption.
+settings.appearance.dateFormat = Date Format
+settings.appearance.dateFormat.description = In which format dates should be displayed in the client.
 settings.checkForPrerelease.description = When checking for client updates, also check for pre-releases.
 settings.prereleaseFeedback.description = Join our Discord channel to give feedback for client pre-releases.
 settings.general.startTab.description = The tab to open when you start the client.
@@ -939,8 +941,6 @@ game.generateMap.blind = Blind
 settings.date.system = Auto
 settings.date.mdy = Month/Day/Year
 settings.date.dmy = Day/Month/Year
-settings.chat.dateFormat = Date Format
-settings.chat.dateFormat.description = In which format the date should be displayed in the chat.
 chat.error.noHashTag = Channel Name must start with \# (Try \#{0})
 chat.message.addError = Could not add message to chat
 game.gameStatus.hosting = Hosting
